requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==3.0.1
psycopg2-binary==2.9.9
openai==1.12.0
httpx==0.24.1
python-dotenv==1.0.0
reportlab==4.0.7
PyPDF2==3.0.1
email-validator==2.1.0
bcrypt==4.1.2
gunicorn==21.2.0
pydub==0.25.1

